move auth routes out of next.js
